+++ record_android/android/src/main/kotlin/com/llfbandit/record/record/RecordConfig.kt
@@ -6,7 +6,7 @@ import android.media.MediaRecorder
 
 class RecordConfig(
     val path: String?,
-    val encoder: String,
+    encoder: String,
     val bitRate: Int,
     val sampleRate: Int,
     numChannels: Int,
@@ -19,21 +19,45 @@ class RecordConfig(
     val manageBluetooth: Boolean = true,
     val audioSource: Int = MediaRecorder.AudioSource.DEFAULT,
     val speakerphone: Boolean = false,
-    val audioManagerMode: Int = AudioManager.MODE_NORMAL
+    val audioManagerMode: Int = AudioManager.MODE_NORMAL,
+    audioInterruption: Int,
 ) {
     val numChannels: Int = 2.coerceAtMost(1.coerceAtLeast(numChannels))
+    val audioInterruption: AudioInterruption = when (audioInterruption) {
+        0 -> AudioInterruption.NONE
+        1 -> AudioInterruption.PAUSE
+        2 -> AudioInterruption.PAUSE_RESUME
+        else -> AudioInterruption.PAUSE
+    }
+
+    val encoder: AudioEncoder = when (encoder) {
+        "aacLc" -> AudioEncoder.AacLc
+        "aacEld" -> AudioEncoder.AacEld
+        "aacHe" -> AudioEncoder.AacHe
+        "amrNb" -> AudioEncoder.AmrNb
+        "amrWb" -> AudioEncoder.AmrWb
+        "flac" -> AudioEncoder.Flac
+        "pcm16bits" -> AudioEncoder.Pcm16bits
+        "opus" -> AudioEncoder.Opus
+        "wav" -> AudioEncoder.Wav
+        else -> AudioEncoder.AacLc
+    }
 }
 
-class AudioEncoder {
-    companion object {
-        const val aacLc = "aacLc"
-        const val aacEld = "aacEld"
-        const val aacHe = "aacHe"
-        const val amrNb = "amrNb"
-        const val amrWb = "amrWb"
-        const val flac = "flac"
-        const val pcm16bits = "pcm16bits"
-        const val opus = "opus"
-        const val wav = "wav"
-    }
+enum class AudioEncoder(val value: String) {
+    AacLc("aacLc"),
+    AacEld("aacEld"),
+    AacHe("aacHe"),
+    AmrNb("amrNb"),
+    AmrWb("amrWb"),
+    Flac("flac"),
+    Pcm16bits("pcm16bits"),
+    Opus("opus"),
+    Wav("wav")
+}
+
+enum class AudioInterruption {
+    NONE,
+    PAUSE,
+    PAUSE_RESUME
 }